AI Дайджест Ассистент/requirements.txt
openai
pandas
streamlit
python-dotenv
transformers 
torch
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2